+++ src/Translators/Lec.ExtProtocol/UnmanagedLibraryLoader.cs
@@ -8,19 +8,16 @@ namespace Lec.ExtProtocol
       private IntPtr _libraryPointer;
       private bool _disposed = false;
 
-      public bool LoadLibrary( string path )
+      public void LoadLibrary( string path )
       {
          _libraryPointer = Kernel32.LoadLibrary( path );
-
-         return _libraryPointer != IntPtr.Zero;
+         if( _libraryPointer == IntPtr.Zero ) throw new Exception( $"Could not load the unmanaged library '{path}'." );
       }
 
       public TDelegate LoadFunction<TDelegate>( string name )
       {
-         if( _libraryPointer == IntPtr.Zero ) throw new InvalidOperationException( $"Could not load the function {name} because no library has been loaded!" );
-
          var addr = Kernel32.GetProcAddress( _libraryPointer, name );
-         if( addr == IntPtr.Zero ) throw new InvalidOperationException( $"Could not find the function pointer {name}!" );
+         if( addr == IntPtr.Zero ) throw new Exception( $"Could not find the function pointer for '{name}'." );
 
          return (TDelegate)(object)Marshal.GetDelegateForFunctionPointer( addr, typeof( TDelegate ) );
       }